requirements/dev.txt
# Everything the developer needs in addition to the production requirements
-r prod.txt

# Testing
pytest==3.7.4
WebTest==2.0.30
factory-boy==2.11.1
coverage
pylint

# Lint and code style
flake8==3.5.0
flake8-blind-except==0.1.1
flake8-debugger==3.1.0
flake8-docstrings==1.3.1
flake8-isort==2.5
flake8-quotes==1.0.0
isort==4.3.4
pep8-naming==0.7.0

nose==1.3.7
halo_aws==0.11.41
boto3==1.9.227

# Flask
connexion >= 2.6.0

Flask-SQLAlchemy==2.3.2


# Migrations
Flask-Migrate==2.2.1

# Forms
Flask-WTF==0.14.2
WTForms==2.2.1

# Deployment
gunicorn>=19.1.1

# Webpack
flask-webpack==0.1.0

# Auth
Flask-Login==0.4.1
Flask-Bcrypt==0.7.1

# Caching
Flask-Caching>=1.0.0

# Debug toolbar
Flask-DebugToolbar==0.10.1

wheel

flask-filter

#####
pytest-icdiff
mypy
pylint
redis
tenacity
